Functional Code to Live Stream
Code works, just need to customize for your directories. Includes a local example rsync stream to laptop, which assumes it has a apache, if not use python to serve from dir for testing. screen recorder is a test mostly to make sure video is captured. Streaming via parec is available in 3 sampling and compression rates. ts live stream can be made to mp3 locally via ffmpeg "cat-ting" all the segments back together into mp3. There are notes, worth a read. Index.html needs to be custom for your server location, uses a js media player. The rsync to laptop or server requires inotifywait to look for the presence of new .ts and .m3u8 files. There is latency, I noticed 45 seconds of it via a browser, no big deal just be aware.